Changed test to work with invalid CA

+++ test/ngrok.postinstall.spec.js
@@ -1,6 +1,5 @@
 const homedir = require('homedir');
 const path = require('path');
-const forge = require('node-forge').pki;
 const fs = require('fs');
 const child_process = require('child_process');
 
@@ -10,15 +9,14 @@ const certpath = `${ngrokPath}/testCert.pem`;
 const npmrcPath = path.join(homedir(), '.npmrc');
 const npmrcBackupPath = path.join(homedir(), '.npmrc.bak');
 
-describe.only('postinstall cert', () => {
+describe('postinstall', () => {
 	before(() => {
 		clearNgrokDirectory();
 		writeTestPemFile();
 		updateNpmrc();
-		console.log(fs.readFileSync(npmrcPath).toString());
 	});
 
-	it('should run using CA without crashing', done => {
+	it('should run using invalid CA without crashing', done => {
 		const postinstall = child_process.fork(postinstallPath, { stdio: 'ignore' });
 		postinstall.on('exit', done);
 	});
@@ -31,7 +29,9 @@ describe.only('postinstall cert', () => {
 
 function restoreFiles() {
 	fs.unlinkSync(npmrcPath);
-	fs.copyFileSync(npmrcBackupPath, npmrcPath);
+	if (fs.existsSync(npmrcBackupPath)) {
+		fs.copyFileSync(npmrcBackupPath, npmrcPath);
+	}
 	fs.unlinkSync(npmrcBackupPath);
 }
 
@@ -41,71 +41,14 @@ function clearNgrokDirectory() {
 }
 
 function writeTestPemFile() {
-	const pem = createPem();
+	const pem = '-junk-to-test-ignore-';
 	fs.writeFileSync(certpath, pem, { flag: 'w' });
 }
 
 function updateNpmrc() {
-	fs.copyFileSync(npmrcPath, npmrcBackupPath);
-	fs.writeFileSync(npmrcPath, `cafile=${certpath}`);
-}
+	if (fs.existsSync(npmrcPath)) {
+		fs.copyFileSync(npmrcPath, npmrcBackupPath);
+	}
 
-function createPem() {
-	const keys = forge.rsa.generateKeyPair(2048);
-	const cert = forge.createCertificate();
-	cert.publicKey = keys.publicKey;
-	cert.serialNumber = '01';
-	cert.validity.notBefore = new Date();
-	cert.validity.notAfter = new Date();
-	cert.validity.notAfter.setHours(cert.validity.notBefore.getHours() + 1);
-	const attrs = [{ name: 'commonName', value: 'localhost' }];
-	cert.setSubject(attrs);
-	cert.setIssuer(attrs);
-	cert.setExtensions([
-		{
-			name: 'basicConstraints',
-			cA: true
-		},
-		{
-			name: 'keyUsage',
-			keyCertSign: true,
-			digitalSignature: true,
-			nonRepudiation: true,
-			keyEncipherment: true,
-			dataEncipherment: true
-		},
-		{
-			name: 'extKeyUsage',
-			serverAuth: true,
-			clientAuth: true,
-			codeSigning: true,
-			emailProtection: true,
-			timeStamping: true
-		},
-		{
-			name: 'nsCertType',
-			client: true,
-			server: true,
-			email: true,
-			objsign: true,
-			sslCA: true,
-			emailCA: true,
-			objCA: true
-		},
-		{
-			name: 'subjectAltName',
-			altNames: [
-				{
-					type: 7, // IP
-					ip: '127.0.0.1'
-				}
-			]
-		},
-		{
-			name: 'subjectKeyIdentifier'
-		}
-	]);
-	cert.sign(keys.privateKey);
-
-	return forge.certificateToPem(cert);
+	fs.writeFileSync(npmrcPath, `cafile=${certpath}`, { flag: 'w' });
 }
